Reconfiguring file structure, added confirmation messages to the "clean" and "quit" buttons, implemented memory optimization by not loading all of the images from the pdf file at once

+++ pdf_sorter_gui.py
@@ -5,12 +5,13 @@ import shutil
 import tkinter as tk
 from tkinter import filedialog
 from tkinter import scrolledtext
+from tkinter import messagebox
 import configparser
 from PIL import Image
 import pytesseract
 from pdf2image import convert_from_path
-from app_scripts.pdf_image_viewer import PdfImageViewer
-from app_scripts.crop_box_selector import PdfCropSelector
+from pdf_image_viewer import PdfImageViewer
+from crop_box_selector import PdfCropSelector
 
 
 class SorterApp:
@@ -30,10 +31,9 @@ class SorterApp:
     Warning: Currently the stacktrace for any errors that occur will only be visible via the terminal
     """
 
-    def __init__(self, root):
+    def __init__(self, root, config_file='config.ini'):
         self.config = configparser.ConfigParser()
-        self.load_config()
-        # self.config.read('config.ini')
+        self.load_config(config_file=config_file)
 
         if sys.platform.startswith('win'):
             pytesseract.pytesseract.tesseract_cmd = self.config.get('SETTINGS', 'tesseract_cmd',
@@ -69,10 +69,10 @@ class SorterApp:
         self.runMenu.add_command(label='OCR', command=self.run_ocr)
         self.runMenu.add_command(label='Image+Text Viewer', command=self.run_main_viewer)
         self.runMenu.add_separator()
-        self.runMenu.add_command(label="Quit", command=self.deactivate)
+        self.runMenu.add_command(label="Quit", command=lambda: self.deactivate(confirmation_box=True))
         self.menuBar.add_cascade(label="Run", menu=self.runMenu)
         self.menuBar.add_command(label="Check", command=self.run_check)
-        self.menuBar.add_command(label="Clean", command=self.output_clean)
+        self.menuBar.add_command(label="Clean", command=lambda: self.output_clean(confirmation_box=True))
         self.menuBar.add_command(label="Clear", command=self.clear_term)
 
         # building right frame
@@ -113,8 +113,8 @@ class SorterApp:
         self.load_box_config()
         self.root.deiconify()
 
-    def load_config(self):
-        self.config.read('config.ini')
+    def load_config(self, config_file='config.ini'):
+        self.config.read(config_file)
 
     def write_config(self):
         with open('config.ini', 'w') as config_file:
@@ -137,9 +137,9 @@ class SorterApp:
         if isinstance(text, str) is False:
             text = str(text)
         if text != '':
-            print('-'+text.expandtabs(self.tab_size))
+            print('-' + text.expandtabs(self.tab_size))
             self.terminal_output.configure(state='normal')
-            self.terminal_output.insert('end', f"{'-'+text.expandtabs(self.tab_size)}\n")
+            self.terminal_output.insert('end', f"{'-' + text.expandtabs(self.tab_size)}\n")
             self.terminal_output.see('end')
             self.terminal_output.configure(state='disabled')
         self.root.update()
@@ -154,9 +154,16 @@ class SorterApp:
         """Starts the mainloop for the tk main window"""
         self.root.mainloop()
 
-    def deactivate(self):
+    def deactivate(self, confirmation_box=False):
         """Destroys the tk main window"""
-        self.root.quit()
+        if confirmation_box:
+            quit_message = tk.messagebox.askquestion('Exit Application',
+                                                     'Are you sure that you want to quit this application?',
+                                                     icon='warning')
+            if quit_message == 'yes':
+                self.root.quit()
+        else:
+            self.root.quit()
 
     def select_input_file(self):
         """Opens a file selection tkinter window for the user to select a pdf file"""
@@ -275,9 +282,13 @@ class SorterApp:
 
         if self.run_check():
             output_dict = {}
-            for image_name in os.listdir(f"{self.output_dir}/images"):
-                image_filename = f"{self.output_dir}/images/{image_name}"
-                crop_filename = f"{self.output_dir}/crops/{image_name}"
+            image_list = os.listdir(f"{self.output_dir}/images")
+            image_list.sort(key=lambda x: x.split('-')[-1].split('.')[0])
+            crop_list = os.listdir(f"{self.output_dir}/crops")
+            crop_list.sort(key=lambda x: x.split('-')[-1].split('.')[0])
+            for index, image_file in enumerate(image_list):
+                image_filename = f"{self.output_dir}/images/{image_file}"
+                crop_filename = f"{self.output_dir}/crops/{crop_list[index-1]}"
                 extracted_text = self.image_extract_text(crop_filename)
                 temp_set = []
                 if extracted_text in output_dict:
@@ -306,29 +317,43 @@ class SorterApp:
         except FileExistsError:
             pass
 
-    def output_clean(self):
+    def output_clean(self, confirmation_box=False):
         """Deletes the folder holding the files that are produced by this program"""
-        try:
-            shutil.rmtree(self.output_dir)
-            self.term_print(f'{self.output_dir} has been deleted')
-        except Exception:
-            self.term_print(f'Error in cleaning {self.output_dir}')
+        if confirmation_box:
+            clean_message = tk.messagebox.askquestion('Clean TEMP Folder',
+                                                      'Are you sure you want to clean the TEMP Folder?',
+                                                      icon='warning')
+            if clean_message == 'yes':
+                try:
+                    shutil.rmtree(self.output_dir)
+                    self.term_print(f'{self.output_dir} has been deleted')
+                except Exception:
+                    self.term_print(f'Error in cleaning {self.output_dir}')
+        else:
+            try:
+                shutil.rmtree(self.output_dir)
+                self.term_print(f'{self.output_dir} has been deleted')
+            except Exception:
+                self.term_print(f'Error in cleaning {self.output_dir}')
 
     def pdf_image_splitter(self, input_file):
         """Splits the pdf file into pages and saves the contents as images"""
         self.create_output_dir()
         self.term_print(input_file.name)
         self.term_print(os.path.basename(input_file.name))
-
-        pdf_file = convert_from_path(input_file.name, dpi=self.config.getint('SETTINGS', 'dpi', fallback=200),
-                                     poppler_path=self.poppler_path)
+        
         self.term_print('pdf found')
+        self.term_print(f"Extracting images from the pages of {input_file.name}")
+        pdf_file_images = convert_from_path(input_file.name, dpi=self.config.getint('SETTINGS', 'dpi', fallback=200),
+                                            poppler_path=self.poppler_path, paths_only=True, fmt="png", thread_count=4,
+                                            output_folder=f"{self.output_dir}/images")
+        self.term_print(f"Created {len(pdf_file_images)} images from pdf file")
 
-        for num, page in enumerate(pdf_file):
-            output_filename = f"{self.output_dir}/images/page_{num + 1}." \
-                              f"{self.config.get('SETTINGS', 'image_type', fallback='png')}"
-            page.save(output_filename)
-            self.term_print(f"Created: {output_filename.split('/')[-1]}")
+        # for num, page in enumerate(pdf_file):
+        #     output_filename = f"{self.output_dir}/images/page_{num + 1}." \
+        #                       f"{self.config.get('SETTINGS', 'image_type', fallback='png')}"
+        #     page.save(output_filename)
+        #     self.term_print(f"Created: {output_filename.split('/')[-1]}")
 
     def crop_image(self, input_file):
         """Crops the given image to the crop box that was selected"""
@@ -336,14 +361,14 @@ class SorterApp:
         self.term_print(f"image {img_name} found")
         file_ext = img_name.split('.')[-1]
 
-        i = img_name.split('_')[-1].split('.')[0]
+        i = img_name.split('-')[-1].split('.')[0]
         page = Image.open(input_file)
         page = page.crop((self.crop_box['start']['x'],
                           self.crop_box['start']['y'],
                           self.crop_box['end']['x'],
                           self.crop_box['end']['y']))
-        page.save(f"{self.output_dir}/crops/page_{i}.{file_ext}")
-        self.term_print(f"image page_{i}.{file_ext} saved")
+        page.save(f"{self.output_dir}/crops/{i}.{file_ext}")
+        self.term_print(f"image {i}.{file_ext} saved")
 
     def image_extract_text(self, input_file):
         """Runs Tesseract-OCR using pytesseract to extract number from the given image and saves the extracted text"""
@@ -359,6 +384,10 @@ class SorterApp:
         return text
 
 
-if __name__ == '__main__':
+def main():
     app = SorterApp(tk.Tk())
     app.activate()
+
+
+if __name__ == '__main__':
+    main()
